change MemoryOptimizer to be an interface, add two different Optimizers
